refactor: extract pure GitHub/date logic into lib
- describeGithubEvent (now typed, no more any) and pickHighlights move
  to lib/github.ts; relativeTime/formatDate to lib/format.ts — pure
  functions, unit-testable without React
- shared useGithubApi hook replaces the duplicated fetch/cancel/fallback
  pattern in ActivityFeed and GithubHighlights
- avatars keyed by profileUrl instead of display name, so renaming a
  recommender can't silently drop their photo
- formatDate pinned to UTC so output doesn't vary by build machine TZ
